refactor: split DB-backed language CRUD into its own controller
language_controller.py contained two concatenated APIRouter
definitions — a merge artifact from the a4i/staging merge that
silently discarded the original static /v1/languages router in
favor of the new DB-backed one, both sharing the same module.

Move the DB-backed CRUD endpoints (GET/POST/PUT/DELETE /languages)
into a new language_management_controller.py, and restore
language_controller.py to its original staging content (static
GET /v1/languages), making the two module concerns unambiguous.

language_registry_controller.py becomes redundant with the restored
language_controller.py (same route, same purpose) and is removed.
