--- platform/app/controllers/language_registry_controller.py
@@ -1,12 +1,0 @@
-from __future__ import annotations
-
-from fastapi import APIRouter
-
-from app.services.language_registry import SUPPORTED_LANGUAGES, Language
-
-router = APIRouter(prefix="/v1", tags=["Languages"])
-
-
-@router.get("/languages", summary="List supported languages (static SDK registry)")
-async def list_supported_languages() -> dict[str, list[Language]]:
-    return {"languages": SUPPORTED_LANGUAGES}
